feat: no-config auto attach to browser
- Removes existing implementation using explicit config options
- Updates the rpc/worker to support reading vite's resolved browser
configuration. This information is used to determine when the
secondary attach config should be started. This information is also
used to set the inspect-brk and browser flags on the vitest CLI when
starting the debug session. This enables zero-config debugging when
the user has started a test debug without interfering with running
without debug. inspect-brk is used to guarantee that breakpoints
early in the test run are hit (otherwise it is possible tests run
before the debugger is attached).
- Adds samples which verify that
this feature works correctly when there are multiple instances
configured and when webdriverio is being used (note: debugging is not
supported for webdriverio, this just makes sure the feature doesn't
break test running).

Only chromium with playwright is supported by the current
implementation.